fix python stubs
